requirements.txt
python-docx
icalendar